backend/requirements.txt
django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
psycopg2-binary>=2.9
python-dotenv>=1.0
requests>=2.31
